+++ CHANGELOG.md
@@ -15,6 +15,7 @@ ENHANCEMENTS:
 * Added a simplified Build Process with a Makefile Target ([#1926](https://github.com/opentofu/opentofu/issues/1926))
 * Ensures that the Makefile adheres to POSIX standards ([#1811](https://github.com/opentofu/opentofu/pull/1928))
 * Added for-each support to providers. ([#300](https://github.com/opentofu/opentofu/issues/300))
+* Added `-pedantic` global option to enable treating warnings as errors. ([#1726](https://github.com/opentofu/opentofu/pull/1726))
 
 BUG FIXES:
 * Ensure that using a sensitive path for templatefile that it doesn't panic([#1801](https://github.com/opentofu/opentofu/issues/1801))